Revert "ortografia"
