+++ .gitignore
@@ -10,4 +10,3 @@ Homestead.json
 Homestead.yaml
 npm-debug.log
 yarn-error.log
-**copy.js**
