+++ resources/js/components/JuridicasPanel.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import TableJuridicas from './TableJuridicas.js';
-import Modal from "react-bootstrap/Modal";
 
 import axios from 'axios';
 
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+
 
 class JuridicasPanel extends Component {
     constructor(props) {
@@ -24,20 +22,13 @@ class JuridicasPanel extends Component {
             estado: '',
             motivo: '',
             desde: '',
-            hasta: '',
-            modal: {
-                show: false,
-                id: ''
-            }
+            hasta: ''
         }
         // bind
         this.getJuridicas = this.getJuridicas.bind(this);
         this.handleChange = this.handleChange.bind(this);
         this.buscar = this.buscar.bind(this);
         this.crearRegistro = this.crearRegistro.bind(this);
-        this.showModal = this.showModal.bind(this);
-        this.handleCerrarModal = this.handleCerrarModal.bind(this);
-        this.handleEliminar = this.handleEliminar.bind(this);
         this.getJuridicas()
     }
 
@@ -59,6 +50,9 @@ class JuridicasPanel extends Component {
                 console.log(err)
             })
     }
+    componentDidMount() {
+
+    }
     exportReport(){
         window.open('exportJuridicas','_blank');
     }
@@ -78,61 +72,9 @@ class JuridicasPanel extends Component {
     crearRegistro(){
         window.open('verJuridica/1/1/c','_blank');
     }
-
-    handleToast(arg,type) {
-        if(type == 'success') {
-            toast.success(arg, {
-                position: "top-right",
-                autoClose: 1500,
-                hideProgressBar: false,
-                closeOnClick: true,
-                pauseOnHover: true,
-                draggable: true,
-                progress: undefined,
-            });
-        }else if(type == 'error') {
-            toast.error(arg, {
-                position: "top-right",
-                autoClose: 1500,
-                hideProgressBar: false,
-                closeOnClick: true,
-                pauseOnHover: true,
-                draggable: true,
-                progress: undefined,
-            });
-        }
-    }   
-
-    handleCerrarModal(){
-        this.setState({
-            modal: {show: false},
-            eliminar: false
-        });
-    }
-
-    showModal(arg, id) {
-        this.setState({modal: {show: arg, id}});
-    }
-
-    handleEliminar() {
-        const id= this.state.modal;
-        const url ="/deleteJuridica"
-        console.log('target eliminar', id);
-        axios.post(url, { datos: id.id })
-        .then(resp => {
-            this.handleToast(resp.data,'success')
-            this.setState({modal: {show: false}})
-        })
-        .catch(err => {
-            this.handleToast(err,'error')
-
-        })
-        this.getJuridicas()
-    }
     render() {
         return (
             <div>
-                <ToastContainer/>
                 <div className="row mt-2">
                     <div className="col-md-12">
                         <div className="card">
@@ -143,13 +85,62 @@ class JuridicasPanel extends Component {
                                         <label htmlFor="">Identificación del usuario</label>
                                         <input className="form-control" name="identificacion" value={this.state.identificacion} onChange={this.handleChange} />
                                     </div>
+                                   {/* <div className="col-md-3">
+                                        <label htmlFor="">Conducta</label>
+                                        <select className="form-control" name="conducta" value={this.state.conducta} onChange={this.handleChange}>
+                                            <option value=""></option>
+                                            {this.state.conductas.map((c) =>
+                                                <option value={c}>{c}</option>
+
+                                            )}
+                                        </select>
+                                            </div>
+                                    <div className="col-md-3">
+                                        <label htmlFor="">Estado</label>
+                                        <select className="form-control" name="estado" value={this.state.estado} onChange={this.handleChange}>
+                                            <option value=""></option>
+                                            {this.state.estados.map((e) =>
+                                                <option value={e}>{e}</option>
+
+                                            )}
+                                        </select>
+                                    </div>
+                                    <div className="col-md-3">
+                                        <label htmlFor="">Motivo</label>
+                                        <select className="form-control" name="motivo" value={this.state.motivo} onChange={this.handleChange}>
+                                            <option value=""></option>
+                                            {this.state.motivos.map((m) =>
+                                                <option value={m}>{m}</option>
+
+                                            )}
+                                        </select>
+                                    </div>*/}
                                     <div className="col-md-6">
                                         <br/>
                                         <button className="btn btn-success" onClick={this.buscar}>Buscar</button>&nbsp;
                                         <button className="btn btn-success" onClick={this.exportReport}>Exportar Datos</button>&nbsp;
                                         <button className="btn btn-success" onClick={this.crearRegistro}>Crear registro</button>
+                                   
                                     </div>
                                 </div>
+                                {/*
+                                <div className="row mt-2">
+                                    <div className="col-md-4">
+                                        <label htmlFor="">Fecha de notificación (desde)</label>
+                                        <input type="date" name="desde" className="form-control form-control-sm" onChange={this.handleChange}/>
+                                    </div>
+                                    <div className="col-md-4">
+                                        <label htmlFor="">Fecha de notificación (hasta)</label>
+                                        <input type="date" name="hasta" className="form-control form-control-sm" onChange={this.handleChange}/>
+                                    </div>
+                                    <div className="col-md-4">
+                                        <br/>
+                                        <button className="btn btn-success" onClick={this.exportReport}>Exportar Datos</button>
+                                    </div>
+                                </div>
+                                <div className="row mt-1">
+                                    <button className="btn btn-success" onClick={this.buscar}>Buscar</button>
+                                            </div>*/}
                             </div>
                         </div>
                     </div>
@@ -159,28 +150,18 @@ class JuridicasPanel extends Component {
                         <div className="card">
                             <div className="card-header bg2 titulo">Listado de Pacientes</div>
                             <div className="card-body texto"></div>
+
+
+
+
                         </div>
                         {this.state.juridicas != '' ? (
-                            <TableJuridicas juridicas={this.state.juridicas} setModal={(arg, id) => this.showModal(arg, id)}/>
+                            <TableJuridicas juridicas={this.state.juridicas} />
                         ) : (
                                 <p>No hay datos</p>
                             )}
                     </div>
                 </div>
-                <Modal show={this.state.modal.show}>
-                    <Modal.Header>Eliminar registro juridico</Modal.Header>
-                    <Modal.Body>
-                        <div className="container">
-                            <div className="row">
-                                <h3>¿Desea eliminar el registro juridico #{this.state.modal.id} ?</h3>
-                            </div>
-                        </div>
-                    </Modal.Body>
-                    <Modal.Footer>
-                        <button className="btn btn-danger btn-sm" onClick={ this.handleEliminar }>Eliminar</button>
-                        <button className="btn btn-primary btn-sm" onClick={ this.handleCerrarModal }>Cerrar</button>
-                    </Modal.Footer>
-                </Modal>
             </div>
         );
     }
