+++ resources/js/containers/Medicos/MedicosSistema.js
@@ -395,7 +395,7 @@ class MedicosSistema extends Component {
                         <div className="form-group">
                             <label htmlFor="codigo">Contraseña</label>
                             <input
-                                pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s)(?=.+([\x20-\x2F]|[\x3A-\x40]|[\x5B-\x60]|[\x7B-\x7E])).{8,16}$"
+                                pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s).{8,20}$"
                                 title="al menos una mayuscula,
                                 al menos una minuscula,
                                 al menos un número,
@@ -416,7 +416,7 @@ class MedicosSistema extends Component {
                         <div className="form-group">
                             <label htmlFor="codigo">Confirmar contraseña</label>
                             <input
-                                pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s)(?=.+([\x20-\x2F]|[\x3A-\x40]|[\x5B-\x60]|[\x7B-\x7E])).{8,16}$"
+                                pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s).{8,20}$"
                                 title="al menos una mayuscula,
                                 al menos una minuscula,
                                 al menos un número,
@@ -512,7 +512,7 @@ class MedicosSistema extends Component {
                                                     <div className="col-md-4">
                                                         <label htmlFor="nombre">Contraseña</label>
                                                         <input
-                                                            pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s)(?=.+([\x20-\x2F]|[\x3A-\x40]|[\x5B-\x60]|[\x7B-\x7E])).{8,16}$"
+                                                            pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s).{8,20}$"
                                                             title="al menos una mayuscula,
                                                             al menos una minuscula,
                                                             al menos un número,
@@ -532,7 +532,7 @@ class MedicosSistema extends Component {
                                                     <div className="col-md-4">
                                                         <label htmlFor="nombre">Confirmar Contraseña</label>
                                                             <input
-                                                            pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s)(?=.+([\x20-\x2F]|[\x3A-\x40]|[\x5B-\x60]|[\x7B-\x7E])).{8,16}$"
+                                                            pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s).{8,20}$"
                                                             title="al menos una mayuscula,
                                                             al menos una minuscula,
                                                             al menos un número,
