+++ routes/web.php
@@ -99,15 +99,14 @@ Route::get('/exportAudits','ExcelController@exportAudits');
 
 //cronicos
 
-Route::get('/cronicos','CronicosController@index')->name('cronicos')->middleware('auth','twofactor');
+Route::get('/cronicos','CronicosController@index')->name('cronicos')->middleware('auth','twofactor');;
 Route::get('/getCronicos','CronicosController@getCronicos');
 Route::post('/buscarCronicos','CronicosController@buscarCronicos');
 Route::post('/buscarCronico','CronicosController@buscarCronico');
 Route::get('/verCronico/{id}/{enable}','CronicosController@verCronico');
 Route::get('/getCronico/{id}','CronicosController@getCronico');
-Route::post('/updateCronico','CronicosController@updateCronico')->middleware('auth','twofactor');
-Route::delete('/deleteCronico/{id}','CronicosController@deleteCronico')->middleware('auth','twofactor');
-Route::post('/addCronico','CronicosController@addCronico')->middleware('auth','twofactor');
+Route::post('/updateCronico','CronicosController@updateCronico');
+
 
 //juridicos
 Route::get('/juridicas','JuridicasController@index')->name('juridicas')->middleware('auth','twofactor');
